Extract `AbstractGeoIpDownloader` for shared concurrency logic (#137660)
In #134223, we modified the concurrency design of the `EnterpriseGeoIpDownloader`. Before that, both `EnterpriseGeoIpDownloader` and `GeoIpDownloader` shared the exact same concurrency implementation. In this PR, we move the improved concurrency design of `EnterpriseGeoIpDownloader` to `AbstractGeoIpDownloader` and make `GeoIpDownloader` extend this class as well. This way, both classes are aligned on the concurrency implementation and will both benefit from any future improvements/modifications.

There is likely more logic in these two classes that can be moved to the abstract class, but I'm leaving that for future work, to keep the scope of this PR limited to the concurrency improvements.

Incidentally, I am optimistic that this will resolve the following test failures. I've been running the GeoIP YAML test suites for thousands of iterations on `main`, but haven't been able to reproduce the failures at all. The reported failure messages are almost exactly the same as the one in #126124, which was addressed by #134223. Therefore, I'm closing these issues for now.
Closes #130681
Closes #133597
Closes #135132
Closes #135158

(cherry picked from commit edb40c564d5ee16cffe58f8ad0d01055ddd98e4c)

# Conflicts:
#	modules/ingest-geoip/src/main/java/org/elasticsearch/ingest/geoip/EnterpriseGeoIpDownloader.java
#	modules/ingest-geoip/src/main/java/org/elasticsearch/ingest/geoip/GeoIpDownloader.java
#	modules/ingest-geoip/src/main/java/org/elasticsearch/ingest/geoip/GeoIpDownloaderTaskExecutor.java
