+++ modules/ingest-geoip/src/main/java/org/elasticsearch/ingest/geoip/GeoIpDownloader.java
@@ -32,10 +32,8 @@ import org.elasticsearch.index.reindex.DeleteByQueryAction;
 import org.elasticsearch.index.reindex.DeleteByQueryRequest;
 import org.elasticsearch.ingest.geoip.GeoIpTaskState.Metadata;
 import org.elasticsearch.ingest.geoip.stats.GeoIpDownloaderStats;
-import org.elasticsearch.persistent.AllocatedPersistentTask;
 import org.elasticsearch.persistent.PersistentTasksCustomMetadata.PersistentTask;
 import org.elasticsearch.tasks.TaskId;
-import org.elasticsearch.threadpool.Scheduler;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.xcontent.XContentParser;
 import org.elasticsearch.xcontent.XContentParserConfiguration;
@@ -56,7 +54,7 @@ import java.util.function.Supplier;
  * Downloads are verified against MD5 checksum provided by the server
  * Current state of all stored databases is stored in cluster state in persistent task state
  */
-public class GeoIpDownloader extends AllocatedPersistentTask {
+public class GeoIpDownloader extends AbstractGeoIpDownloader {
 
     private static final Logger logger = LogManager.getLogger(GeoIpDownloader.class);
 
@@ -79,14 +77,11 @@ public class GeoIpDownloader extends AllocatedPersistentTask {
     private final Client client;
     private final HttpClient httpClient;
     private final ClusterService clusterService;
-    private final ThreadPool threadPool;
     private final String endpoint;
 
     // visible for testing
     protected volatile GeoIpTaskState state;
-    private volatile Scheduler.ScheduledCancellable scheduled;
     private volatile GeoIpDownloaderStats stats = GeoIpDownloaderStats.EMPTY;
-    private final Supplier<TimeValue> pollIntervalSupplier;
     private final Supplier<Boolean> eagerDownloadSupplier;
     /*
      * This variable tells us whether we have at least one pipeline with a geoip processor. If there are no geoip processors then we do
@@ -111,13 +106,11 @@ public class GeoIpDownloader extends AllocatedPersistentTask {
         Supplier<Boolean> eagerDownloadSupplier,
         Supplier<Boolean> atLeastOneGeoipProcessorSupplier
     ) {
-        super(id, type, action, description, parentTask, headers);
+        super(id, type, action, description, parentTask, headers, threadPool, pollIntervalSupplier);
         this.client = client;
         this.httpClient = httpClient;
         this.clusterService = clusterService;
-        this.threadPool = threadPool;
         this.endpoint = ENDPOINT_SETTING.get(settings);
-        this.pollIntervalSupplier = pollIntervalSupplier;
         this.eagerDownloadSupplier = eagerDownloadSupplier;
         this.atLeastOneGeoipProcessorSupplier = atLeastOneGeoipProcessorSupplier;
     }
@@ -286,16 +279,15 @@ public class GeoIpDownloader extends AllocatedPersistentTask {
         return buf;
     }
 
-    /**
-     * Downloads the geoip databases now, and schedules them to be downloaded again after pollInterval.
-     */
+    @Override
     void runDownloader() {
-        // by the time we reach here, the state will never be null
-        assert state != null;
-
         if (isCancelled() || isCompleted()) {
+            logger.debug("Not running downloader because task is cancelled or completed");
             return;
         }
+        // by the time we reach here, the state will never be null
+        assert this.state != null : "this.state is null. You need to call setState() before calling runDownloader()";
+
         try {
             updateDatabases();
         } catch (Exception e) {
@@ -306,22 +298,6 @@ public class GeoIpDownloader extends AllocatedPersistentTask {
             cleanDatabases();
         } catch (Exception e) {
             logger.error("exception during geoip databases cleanup", e);
-        }
-        scheduleNextRun(pollIntervalSupplier.get());
-    }
-
-    /**
-     * This method requests that the downloader be rescheduled to run immediately (presumably because a dynamic property supplied by
-     * pollIntervalSupplier or eagerDownloadSupplier has changed, or a pipeline with a geoip processor has been added). This method does
-     * nothing if this task is cancelled, completed, or has not yet been scheduled to run for the first time. It cancels any existing
-     * scheduled run.
-     */
-    public void requestReschedule() {
-        if (isCancelled() || isCompleted()) {
-            return;
-        }
-        if (scheduled != null && scheduled.cancel()) {
-            scheduleNextRun(TimeValue.ZERO);
         }
     }
 
@@ -343,22 +319,7 @@ public class GeoIpDownloader extends AllocatedPersistentTask {
     }
 
     @Override
-    protected void onCancelled() {
-        if (scheduled != null) {
-            scheduled.cancel();
-        }
-        markAsCompleted();
-    }
-
-    @Override
     public GeoIpDownloaderStats getStatus() {
         return isCancelled() || isCompleted() ? null : stats;
     }
-
-    private void scheduleNextRun(TimeValue time) {
-        if (threadPool.scheduler().isShutdown() == false) {
-            scheduled = threadPool.schedule(this::runDownloader, time, threadPool.generic());
-        }
-    }
-
 }
